made changePassword Component
