+++ client/src/pages/student/StudentDashboard.jsx
@@ -11,11 +11,10 @@ import "../../styles/webPage.css";
 import { JWT_KEY, useLocalState } from "../../hooks/useLocalStorage";
 import { apiChangePassword, apiCheckToken } from "../../lib/api";
 import { decodeJWT } from "../../utils/decodeJWT";
+import ChangePassword from "../../components/changePassword";
 
 const StudentDash = () => {
   const [jwt, setJwt] = useLocalState("", JWT_KEY);
-  const [currentPassword, setCurrentPassword] = useState("");
-  const [newPassword, setNewPassword] = useState("");
   const [isOpen, setIsOpen] = useState(false);
 
   const navigate = useNavigate();
@@ -23,27 +22,6 @@ const StudentDash = () => {
   const handleLogout = () => {
     localStorage.removeItem(JWT_KEY);
     navigate("/");
-  };
-
-  const handleChangePassword = async (event) => {
-    event.preventDefault();
-
-    if (currentPassword === newPassword) {
-      alert("passwords must be different");
-      return;
-    }
-
-    const response = await apiChangePassword(currentPassword, newPassword, jwt);
-
-    if (response.error) {
-      alert(`${response.error}`);
-    } else {
-      alert(`${response.success}`);
-
-      setCurrentPassword("");
-      setNewPassword("");
-      setIsOpen(false);
-    }
   };
 
   useEffect(() => {
@@ -88,61 +66,17 @@ const StudentDash = () => {
           Change Password
         </button>
 
-        <Dialog
-          open={isOpen}
-          onClose={() => setIsOpen(false)}
-          className="dialog-pop"
-        >
-          <div className="dialog-pop-back">
-            <div className="pop-panel">
-              <DialogPanel>
-                <DialogTitle className="font-bold">Change Password</DialogTitle>
-                <Description>This will update your password.</Description>
-                <p>
-                  Please enter your current password to confirm and your new
-                  password.
-                </p>
-
-                <form onSubmit={handleChangePassword} className="password-form">
-                  <div className="form-group">
-                    <label htmlFor="currentPassword">Current Password</label>
-                    <input
-                      type="password"
-                      id="currentPassword"
-                      value={currentPassword}
-                      onChange={(e) => setCurrentPassword(e.target.value)}
-                      required
-                    />
-                  </div>
-
-                  <div className="form-group">
-                    <label htmlFor="newPassword">New Password</label>
-                    <input
-                      type="password"
-                      id="newPassword"
-                      value={newPassword}
-                      onChange={(e) => setNewPassword(e.target.value)}
-                      required
-                    />
-                  </div>
-
-                  <div className="form-actions">
-                    <button
-                      className="form-button"
-                      type="button"
-                      onClick={() => setIsOpen(false)}
-                    >
-                      Cancel
-                    </button>
-                    <button className="form-button" type="submit">
-                      Confirm
-                    </button>
-                  </div>
-                </form>
-              </DialogPanel>
-            </div>
+        <Dialog open={isOpen} onClose={() => setIsOpen(false)} className="dialog-pop">
+        <div className="dialog-pop-back">
+          <div className="pop-panel">
+            <DialogPanel>
+              <DialogTitle className="font-bold">Change Password</DialogTitle>
+              <Description>This will update your password.</Description>
+              <ChangePassword jwt={jwt} setIsOpen={setIsOpen} />
+            </DialogPanel>
           </div>
-        </Dialog>
+        </div>
+      </Dialog>
       </div>
     </div>
   );
